Fix Server.respond to match spec (where possible)
Unfortunately, the spec is somewhat vague and inconsistent when it
comes to handling errors in notifications. I chose to follow the spirit
of the spec. To wit:

Server.respond will return an Error object if it is certain that the
request was a notification. Otherwise, it returns a JSON-RPC 2.0
error response.

Server.respond now returns null on successful handling of a
notification. Otherwise, it returns a JSON-RPC 2.0 result response.