Implement Client
Create a client like so:
    var client = new Client({ server: serverInstance });

The server config option is mandatory and must be an instance of Server (or an
object that implements the same interface).

Client instances provide two methods: request and notify. They can be called
with a config object:
    client.request({
        method: 'echo',
        params: [1, 2, 3],
        callback: function (error, result) { ... }
    });
or with an argument list, where the last argument may be a callback:
    client.request('echo', 1, 2, 3, function (error, result) { ... });

Only the config object form supports named params.

Callbacks will be passed two arguments: the error and the result. One of these
will be null. Client.notify ignores any callback.