Add support for asynchronous RPC methods
Previously, methods provided by the server had to return a response
synchronously. This was a bummer.

RPC methods are now called with a context that provides a `this.async` method,
which returns a node-style callback. If the RPC method calls that method, then
the server will wait for the callback to be invoked before responding to the
request.

This change necessitated adding an optional `callback` argument to
`Server.respond`.  Passing a callback is now the only way to obtain the result
of the RPC. Thus, it may be omitted for notifications, but must be provided for
other requests.

Also remove some tests that asserted notification behavior that was not
compliant with the spec (rather than refactoring them to support this patch).

[fix #13]